0068-Kavishka
made shop item slider as a component
